#115 TTSTool 삭제 및 SoundManager로 병합
