+++ Oculo/Utilities/HealthkitManager.swift
@@ -200,5 +200,3 @@ class HealthKitManager {
 //    }
 }
 
-// HealthKitManager를 다양한 곳에서 사용하기 때문에 전역 변수로 따로 빼두었습니다.
-var healthKitManager = HealthKitManager()
